cleanup schema fetching
