+++ src/reducers/content.js
@@ -1,13 +1,8 @@
-import {
-  CONTENT_LOADED,
-  CONTENT_DELETE,
-  SCHEMA_LOADED,
-} from '../actions/content';
+import { CONTENT_LOADED, CONTENT_DELETE } from '../actions/content';
 
 export const initialState = {
   contentList: [],
   links: {},
-  schema: {},
 };
 
 export default (state = initialState, action) => {
@@ -41,16 +36,6 @@ export default (state = initialState, action) => {
         contentList: state.contentList.filter(
           content => content.id !== action.payload.content.id,
         ),
-      };
-    }
-
-    case SCHEMA_LOADED: {
-      return {
-        ...state,
-        schema: {
-          [`${action.payload.entityTypeId}--${action.payload.bundle}`]: action
-            .payload.schema,
-        },
       };
     }
